+++ packages/calendar/src/components/calendar-types/date-range-calendar/DateRangeCalendar.stories.tsx
@@ -1,3 +1,4 @@
+import { Box } from "@stenajs-webui/core";
 import { Store, withState } from "@dump247/storybook-state";
 import {
   DateRangeCalendar,
@@ -6,7 +7,10 @@ import {
   setDayStateValue,
   useDateRangeCalendarState,
   TranslationProvider,
-  getTranslations
+  getTranslations,
+  travelCalendarTheme,
+  dayHighlightSelect,
+  ExtraDayContentProps
 } from "@stenajs-webui/calendar";
 import { storiesOf } from "@storybook/react";
 import { addDays } from "date-fns";
@@ -162,6 +166,69 @@ storiesOf("calendar/Calendar/DateRangeCalendar", module)
         />
       </TranslationProvider>
     )),
+    {
+      notes: { markdown }
+    }
+  )
+  .add(
+    "travel",
+    withState<State>({
+      startDate: undefined,
+      endDate: undefined,
+      focusedInput: "startDate"
+    })(({ store }: { store: Store<State> }) => {
+      const renderDayContent: React.ComponentType<ExtraDayContentProps<{}>> = ({
+        dayState
+      }) =>
+        dayHighlightSelect<JSX.Element | null>(
+          dayState,
+          [],
+          ["today", "enabled"],
+          [
+            <Box
+              position={"absolute"}
+              top={"3px"}
+              left={"3px"}
+              style={{
+                backgroundColor: "#ea143d",
+                width: "6px",
+                height: "6px",
+                borderRadius: "50%"
+              }}
+            />,
+            null
+          ],
+          <Box
+            position={"absolute"}
+            top={0}
+            left={0}
+            bottom={0}
+            right={0}
+            background="#949494"
+            width="70%"
+            height="1px"
+            style={{
+              margin: "auto",
+              transform: "rotate(-45deg)"
+            }}
+          />
+        ) as JSX.Element;
+      return (
+        <DateRangeCalendar
+          highlightToday
+          startDate={store.state.startDate}
+          endDate={store.state.endDate}
+          focusedInput={store.state.focusedInput}
+          setStartDate={startDate => store.set({ startDate })}
+          setEndDate={endDate => store.set({ endDate })}
+          setFocusedInput={focusedInput => store.set({ focusedInput })}
+          defaultHighlights={["disabled"]}
+          theme={travelCalendarTheme}
+          statePerMonth={statePerMonthWithTwoWeeksEnabled}
+          extraDayContent={renderDayContent}
+        />
+      );
+    }),
     {
       notes: { markdown }
     }